pyproject.toml
[project]
name = "romanesco"
version = "1.3.0"
description = "Expenses reporting system"
urls = {repository = "https://github.com/rubenseyer/romanesco"}
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
  "Flask ~= 2.2.5",
  "pdfminer.six >= 20221105"
]

[project.optional-dependencies]
bjoern = ["bjoern"]
botccoli = [
  "selenium ~= 4.8.0",
  "requests ~= 2.29.0",
  "cryptography ~= 40.0.1"
]
sqlite = ["apsw ~= 3.41.0.0"]
postgresql = ["psycopg[c] ~= 3.1.9"]

[project.scripts]
romanesco = "romanesco.__main__:main"
botccoli = "romanesco.botccoli.__main__:main"

[tool.setuptools]
include-package-data = false
package-data = {romanesco = ["static/*", "templates/*", "xdb/*.sql"]}

[tool.setuptools.packages.find]
where = [""]
include = ["romanesco*"]
namespaces = false

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
